==> Auto_Innate_GDA/0.5_ImmuneDB_Network_Schematic.Rmd <==
---
title: "0.5_ImmuneDB_Network_Schematic"
author: "Ursula Widocki"
date: '2022-06-24'
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

```{r}
library(tidyverse)
library(dplyr)
library(magrittr)
library(igraph)
```

```{r}
# make edgelist for graph

df <- data.frame(Node.1 = c(1, 1, 1, 1, 1, 2, 2, 2, 3, 3, 3, 3, 4, 4, 4, 5, 5, 5, 5, 6, 6, 7, 8, 8, 9, 10, 10, 10, 11, 11, 11, 12, 12, 13, 13, 14, 14, 14, 14, 15, 15, 15, 15, 16, 16, 17, 17, 17, 17, 18, 18, 19, 19, 20, 20, 21, 22, 22, 22, 23, 24, 24, 24, 26, 26, 26, 26, 27, 28, 29, 29, 30, 30, 31, 31, 32, 32, 33, 34),
                 Node.2 = c(2, 3, 4, 6, 7, 3, 4, 11, 4, 5, 7, 9, 5, 11, 12, 9, 12, 15, 16, 9, 10, 8, 9, 29, 17, 15, 17, 29, 12, 14, 13, 13, 18, 14, 20, 16, 18, 19, 20, 16, 17, 18, 26, 17, 18, 22, 26, 27, 29, 22, 23, 20, 23, 21, 25, 23, 23, 24, 27, 25, 25, 27, 28, 27, 28, 29, 34, 2, 34, 30, 34,31, 34, 33, 35, 34, 35, 34, 35))

g <- graph_from_data_frame(df, directed = F)
```

```{r}
# add which nodes belong to which databases
GO_nodes <- c(1,2,3,4,5,6,7,8,9,10,11,12,13,14,15,16,17)
Innate_nodes <- c(6,7,8,9,10,15,16,17,26,27,28,29,30,31,32,33,34,35)
KEGG_nodes <- c(15,16,17,11,12,13,14,26,27,28,18,19,20,21,22,23,24,25)

# add disease names as  feature
V(g)$GO <- ifelse(V(g)$name %in% GO_nodes, 1, 0)
V(g)$Innate <- ifelse(V(g)$name %in% Innate_nodes, 1, 0)
V(g)$KEGG <- ifelse(V(g)$name %in% KEGG_nodes, 1, 0)

vs = V(g)
values <- lapply(1:length(vs), function(x) c(V(g)$GO[x], V(g)$Innate[x], V(g)$KEGG[x]))

E(g)$weight <- edge_betweenness(g) %>% CoDiNA::normalize()
coord <- g %>%
  layout_with_fr(., weights = E(.)$weight *2.5) %>%
  igraph::norm_coords()

```

```{r}
# plot graph

plot(g,
     vertex.shape = "pie",
     vertex.pie=values,
     vertex.pie.color = list(c("#820933", "#3ABEFF", "#26FFE6")),
     vertex.label = NA,
     vertex.size = 5,
     vertex.frame.color = NA,
     
     edge.curved = 0.1)
title("Immune Database Schematic Example")

legend(1, -1, legend = c("Database",
"GO","InnateDB","KEGG"), 
  bty = "n",
  ncol = 1,
  lwd = 10,
  horiz = F,
  lty = c(NA, NA, NA, NA),
  pch = c(NA, 20, 20, 20),
  col=c(NA, "#820933", "#3ABEFF", "#26FFE6"),
  cex = 0.8 # font size
)

```
