==> Auto_Innate_GDA/2.3_newData2_Disease_Immune_cell_LCC.Rmd <==
---
title: "2.3_newData2_Disease_Immune_cell_LCC"
author: "Ursula Widocki"
date: "4/3/2022"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

```{r}
library(tidyverse)
library(data.table)
library(magrittr)
library(igraph)
library(ggraph)
library(NetSci)
library(patchwork)
library(Cairo)

`%ni%` <- Negate(`%in%`)

source("/Users/ursulawidocki/Desktop/BarabasiLab/NetMedTools.R")
```

# Read in data
```{r}
## Read in data
# new PPI
ppi <- fread('/Users/ursulawidocki/Desktop/BarabasiLab/Data/PPI_2022_04042022.csv')
ppi_df <- ppi[,c("HGNC_Symbol.1", "HGNC_Symbol.2")]

ppi_df = ppi_df[!(!is.na(ppi_df$HGNC_Symbol.2) & ppi_df$HGNC_Symbol.2 ==""), ]
ppi_df = ppi_df[!(!is.na(ppi_df$HGNC_Symbol.1) & ppi_df$HGNC_Symbol.1 ==""), ]

ppi_df$value = 1
ppi_df = ppi_df %>% unique()
ppi_g <- igraph::graph_from_data_frame(ppi_df, directed = F)
ppi_g <- simplify(ppi_g)

ppi_degree_df = data.frame(Degree = igraph::degree(ppi_g)) %>%
  mutate(Gene = row.names(.))

# new GDA
GDA_data = fread("/Users/ursulawidocki/Desktop/BarabasiLab/Data/GDA_auto_new2.tsv") # has isAutoimmune col
# select disease that have some form of Strong or Weak evidence and have more than 5 disease genes
GDA_auto = GDA_data %>%
  filter(isAutoimmune == 1) %>%
  filter(Strong > 0 | 
           Weak > 0) %>%
  filter(HGNC_Symbol %in% V(ppi_g)$name) %>% 
  group_by(NewName) %>%
  mutate(Total_Genes = n()) %>%
  filter(Total_Genes > 5) %>%
  dplyr::select(NewName, HGNC_Symbol) %>%
  unique()

## Read in autoimmune disease data to get dis with stat signif module
autoimmune_LCC_df = read.table("/Users/ursulawidocki/Desktop/BarabasiLab/Data/AutoDis_LCCs_new2.tsv", 
                               sep = "\t", stringsAsFactors = F, header = T)
rownames(autoimmune_LCC_df) <- autoimmune_LCC_df$Disease

autoimmune_LCC_df <- autoimmune_LCC_df %>% 
  mutate(padj = p.adjust(p_val, method = "fdr")) %>%
  mutate(signif = ifelse(padj < 0.05, "padj < 0.05", "padj > 0.05"))

signif_dis <- autoimmune_LCC_df %>%
  filter(padj < 0.05) %>%
  pull(Disease)

# Immune cell data
immune_cell_df <- fread("/Users/ursulawidocki/Desktop/BarabasiLab/Data/HPA_rna_blood_cell.tsv")
immune_cell_df %<>% dplyr::select(-Gene) %>% dplyr::select(`Gene name`, `Blood cell`, TPM)
summary_immune_cell = immune_cell_df %>% 
  group_by(`Blood cell`, `Gene name`) %>%
  summarise(sum = sum(TPM)) %>% 
  mutate(Q10 = quantile(sum, 0.1),
         Q25 = quantile(sum, 0.25),
         Q50 = quantile(sum, 0.50), 
         Q75 = quantile(sum, 0.75), 
         Q90 = quantile(sum, 0.90))

innate_cell <- c("basophil", "eosinphil", "neutrophil", "classical monocyte", "non-classical monocyte", "intermediate monocyte", "myeloid DC", "plasmacytoid DC", "NK-cell")
adaptive_cell <- c("memory B-cell", "naive B-cell", "gdT-cell", "MAIT T-cell", "T-reg", "naive CD4 T-cell", "memory CD4 T-cell", "naive CD8 T-cell", "memory CD8 T-cell")
```


# calc disease immune cell LCCs
```{r}
## Immune LCCs of diseases

immune_cell_info <- summary_immune_cell %>% 
  filter(sum >= Q50) %>%
  select(`Blood cell`,`Gene name`)
cells = unique(summary_immune_cell$`Blood cell`)
reps = 1000
pb <- txtProgressBar(min = 0, max = length(cells), style = 3)
final_LCCs = list()

for(i in 1:length(cells)){
  setTxtProgressBar(pb, i)
  cell =  cells[i]
  genes_in_cell = immune_cell_info %>% 
    filter(`Blood cell` == cell) %>% 
    filter(`Gene name` %in% V(ppi_g)$name) %>% 
    pull(`Gene name`)
  
  n_genes_cell = length(genes_in_cell)
  
  cell_ppi = induced_subgraph(ppi_g, genes_in_cell)
  
  out = list()
  for(j in 1:length(signif_dis)){
    genes_disease = GDA_auto %>%
      filter(NewName %in% signif_dis[j]) %>%
      filter(HGNC_Symbol %in% V(cell_ppi)$name)
    n_genes_disease_cell  = nrow(genes_disease)
    
    if(n_genes_disease_cell > 0){
      LCC_disease_cell= NetSci::LCC_Significance(N = reps, 
                                                    Targets = genes_disease$HGNC_Symbol, 
                                                    G = cell_ppi,
                                                    bins = 1
      )
      
      # store data
      tmp = data.frame(Disease_Name = signif_dis[j], 
                       Cell = cell, 
                       Cell_Genes = n_genes_cell, 
                       Disease_Cell_Genes = n_genes_disease_cell, 
                       
                       LCC_size = LCC_disease_cell$LCC, 
                       LCC_p = LCC_disease_cell$emp_p)
      
      out[[j]] = tmp
    } else {
      
      tmp = data.frame(Disease_Name = signif_dis[j], 
                       Cell = cell, 
                       Cell_Genes = n_genes_cell, 
                       Disease_Cell_Genes = n_genes_disease_cell, 
                       
                       LCC_size = 0, 
                       LCC_p = 1)
      
      out[[j]] = tmp
    }
  }  
  final_LCCs[[i]] = out %<>% bind_rows()
}

final_LCCs %<>% bind_rows()
```
# Adjust p-values
```{r}
data <- final_LCCs

get_total_genes <- autoimmune_LCC_df[,c("Disease","num_genes")]
colnames(get_total_genes)[1] <- "Disease_Name"

data <- full_join(data, get_total_genes)

data %<>%
  mutate(padj = p.adjust(LCC_p, method = "fdr")) %>%
  mutate(Significance = ifelse(padj < 0.05, "padj < 0.05", "padj > 0.05")) %>%
  mutate(rLCC = (LCC_size/Disease_Cell_Genes)*100) %>%
  mutate(rLCC_50 = ifelse(rLCC > 50, "big rLCC", "small rLCC")) %>%
  mutate(shape = ifelse(padj < 0.05, 1, 4))

```
# write file
```{r}
write.table(data, "/Users/ursulawidocki/Desktop/BarabasiLab/Data/Disease_ImmuneCell_TPM_LCCs_newData2.tsv", row.names = F, col.names = T, sep = "\t")
```

# read file just in case
```{r}
data <- fread("/Users/ursulawidocki/Desktop/BarabasiLab/Data/Disease_ImmuneCell_TPM_LCCs_newData2.tsv")
```

# Make the plot
```{r}
cell_order <- c("T-reg", "naive CD8 T-cell", "naive CD4 T-cell", "memory CD8 T-cell", "memory CD4 T-cell", "MAIT T-cell", "gdT-cell", "naive B-cell", "memory B-cell", "NK-cell",
                "plasmacytoid DC", "myeloid DC", "classical monocyte", "non-classical monocyte", "intermediate monocyte", "basophil","eosinophil", "neutrophil")

data %<>% filter(Cell != "total PBMC")

data$Cell <- as.character(data$Cell)
data$Cell <- factor(data$Cell, levels=cell_order)

p1 <- data %>%
  filter(Disease_Name %in% signif_dis) %>%
  ggplot() +
  aes(x = Disease_Name, 
      y = Cell, 
      color = rLCC_50,
      size = rLCC,
      shape = Significance) +
  geom_point() +
  geom_rect(aes(xmin = 0.6, xmax = Inf, ymin = 0.6, ymax = 9.5), fill = "#F2D7D5", alpha = 0.05, color = NA) +
  geom_rect(aes(xmin = 0.6, xmax = Inf, ymin = 9.5, ymax = Inf), fill = "#D6EAF8", alpha = 0.2, color = NA) +
  geom_point() +
  scale_color_manual(values = c("big rLCC" = "#007CBE", "small rLCC" = "grey75")) +
  scale_shape_manual(values = c(16, 1)) +
  scale_size(range = c(0.1, 5)) + 
  labs(x = "Autoimmune Disease") +
  theme_minimal() +
  theme(axis.text.x = element_text(angle = 90), 
        legend.position = "bottom", 
        panel.grid.major = element_blank(), 
        panel.grid.minor = element_blank(),
        text = element_text(size = 20))

# Disease bar plot
temp <- data %>%
  filter(Disease_Name %in% signif_dis) %>% 
  filter(padj < 0.05) %>% # uveitis dropped here
  group_by(Disease_Name) %>% 
  dplyr::summarise(n=n())
# add rows for diseases with no modules in cells
temp <- rbind(temp, c("narcolepsy", 0)) %>% as.data.frame()
temp$n <- as.integer(temp$n)

p3 <- temp %>%
  ggplot() +
  aes(x = Disease_Name, y = n) +
  geom_bar(fill = "#122265", stat = "identity") + #position = position_dodge(preserve = "single")
  theme_minimal() +
  theme(axis.title.x=element_blank(),
        axis.text.x=element_blank(),
        axis.ticks.x=element_blank()) +
  labs(y = "LCC count")


# Immune cell bar plot
p2 <- data %>%
  filter(Disease_Name %in% signif_dis) %>% 
  filter(padj < 0.05) %>% 
  group_by(Cell) %>% 
  dplyr::summarise(n=n()) %>%
  ggplot() +
  aes(x = Cell, weight = n) +
  geom_bar(fill = "#122265") + # "#333366"
  labs(y = "LCC count") +
  theme_minimal() +
  theme(axis.title.y = element_blank(),
        axis.text.y = element_blank(),
        axis.ticks.y = element_blank()) +
  coord_flip()


CairoPDF("~/Desktop/BarabasiLab/Auto_Innate_GDA/Figures/Disease_ImmuneCell_TPM_LCC_newData2_Q50.pdf", width = 15, height = 10)
# patchwork things together

((p3 + plot_spacer() + plot_layout(widths = c(5,1))) / ((p1 + p2) + plot_layout(widths = c(5,1))))+ plot_layout(heights = c(1,6))

dev.off()

```



```{r}
temp <- data %>%
  filter(Disease_Name %in% signif_dis) %>% 
  filter(padj < 0.05) %>%
  filter(Cell != "total PBMC") %>%
  mutate(ImmComp = ifelse(Cell %in% adaptive_cell, "adaptive", ifelse(Cell %in% innate_cell, "innate", NA))) %>%
  #group_by(Cell) %>%
  group_by(ImmComp) %>% 
  dplyr::summarise(n=n())
```




# Examine autoimmune disease-immune cell comorbidity
```{r}
# Examine autoimmune disease- immune cell comorbidity
cells <- unique(data$Cell)
diseases <- signif_dis
#col_labels <- c("Disease1", "Disease2", "NumCellsDis1", "NumCellsDis2", "CellComorbidity")
#dis_cell_comorb_df <- data.frame(matrix(ncol = length(col_labels), nrow = 0))
#colnames(dis_cell_comorb_df) <- col_labels

k = 0
dis_cell_comorb_df = list()
for(i in 1:(length(cells)-1)){
  # get tissues where disease has a module
  dis1 = diseases[i]
  dis1_df <- data %>%
    filter(Disease_Name == dis1) %>%
    filter(padj < 0.05) %>%
    select(Disease_Name, Cell)
  
  for(j in (i+1):(length(diseases))){
    # get tissues where disease has a module
    k=k+1
    dis2 = diseases[j]
    dis2_df <- data %>%
      filter(Disease_Name == dis2) %>%
      filter(padj < 0.05) %>%
      select(Disease_Name, Cell)
    
    # calc tissue comorbidity score
    cell_dis1 <- unique(dis1_df$Cell) # num tissues dis1
    cell_dis2 <- unique(dis2_df$Cell) # num tissues dis2
    
    inter <- intersect(cell_dis1, cell_dis2) # num cells of both diseases
    total_cell <- unique(union(cell_dis1, cell_dis2)) # total tissues between them

    com_score = length(inter) / length(total_cell) # fraction of common and total tissues
    
    num_innate <- length(inter %in% innate_cell)
    num_adaptive <- length(inter %in% adaptive_cell)
    
    dis_cell_comorb_df[[k]] <- data.frame(Disease1 = dis1, Disease2 = dis2, 
                                          NumCellsDis1 = length(cell_dis1), NumCellsDis2 = length(cell_dis2), CellComorbidity = com_score, 
                                          NumInnateCells = num_innate, NumAdaptiveCells = num_adaptive) # save this as a df, not vect
    
    
  }
}
dis_cell_comorb_df %<>% bind_rows()
```

# graph the disease-immune cell comorbidity above
```{r}
# graph the above

df_for_graph <- dis_cell_comorb_df %>%
  filter(CellComorbidity != 0) %>%
  pivot_longer(cols = c("NumInnateCells", "NumAdaptiveCells"), names_to = "ImmuneType") %>%
  mutate(color = ifelse(ImmuneType == "NumInnateCells", "#8EEDF7", "#F08B8B")) %>%
  select(Disease1, Disease2, color)

cell_comb_g <- graph_from_data_frame(df_for_graph, directed = F)
V(cell_comb_g)$degree <- degree(cell_comb_g)

# I want to plot it such that the higher the comorbidity score, the closer the nodes are together
#set.seed(1255)
coord <- cell_comb_g %>%
  layout_with_drl(., weights = E(.)$CellComorbidity * 2.5) %>%
  igraph::norm_coords()

CairoPDF("/Users/ursulawidocki/Desktop/BarabasiLab/Auto_Innate_GDA/Figures/Autoimmune_Disease_Graph_Innate_Adaptive_Cells.pdf", height = 10, width = 12)
plot(x = cell_comb_g,
     vertex.frame.color = NA,
     vertex.color = "#127475",
     vertex.label.cex = V(cell_comb_g)$degree / 12,
     vertex.label.color = "black",
     vertex.label.degree = -pi/2, # how to move labels?
     vertex.label.dist = 0.1,
     vertex.label.family = "Helvetica",
     vertex.label.font = 2,
     vertex.size = 10,
     vertex.label.y = 3,
     
     edge.color = E(cell_comb_g)$color,
     edge.curved = 0.3,
     
     #layout = coord
     )
#legend("topleft", legend=levels(E(cell_comb_g)$ImmuneType),)

dev.off()

```





# Q10
```{r}
immune_cell_info <- summary_immune_cell %>% 
  filter(sum >= Q10) %>%
  select(`Blood cell`,`Gene name`)
cells = unique(summary_immune_cell$`Blood cell`)
reps = 1000
pb <- txtProgressBar(min = 0, max = length(cells), style = 3)
final_LCCs = list()

for(i in 1:length(cells)){
  setTxtProgressBar(pb, i)
  cell =  cells[i]
  genes_in_cell = immune_cell_info %>% 
    filter(`Blood cell` == cell) %>% 
    filter(`Gene name` %in% V(ppi_g)$name) %>% 
    pull(`Gene name`)
  
  n_genes_cell = length(genes_in_cell)
  
  cell_ppi = induced_subgraph(ppi_g, genes_in_cell)
  
  out = list()
  for(j in 1:length(signif_dis)){
    genes_disease = GDA_auto %>%
      filter(NewName %in% signif_dis[j]) %>%
      filter(HGNC_Symbol %in% V(cell_ppi)$name)
    n_genes_disease_cell  = nrow(genes_disease)
    
    if(n_genes_disease_cell > 0){
      LCC_disease_cell= NetSci::LCC_Significance(N = reps, 
                                                    Targets = genes_disease$HGNC_Symbol, 
                                                    G = cell_ppi,
                                                    bins = 1
      )
      
      # store data
      tmp = data.frame(Disease_Name = signif_dis[j], 
                       Cell = cell, 
                       Cell_Genes = n_genes_cell, 
                       Disease_Cell_Genes = n_genes_disease_cell, 
                       
                       LCC_size = LCC_disease_cell$LCC, 
                       LCC_p = LCC_disease_cell$emp_p)
      
      out[[j]] = tmp
    } else {
      
      tmp = data.frame(Disease_Name = signif_dis[j], 
                       Cell = cell, 
                       Cell_Genes = n_genes_cell, 
                       Disease_Cell_Genes = n_genes_disease_cell, 
                       
                       LCC_size = 0, 
                       LCC_p = 1)
      
      out[[j]] = tmp
    }
  }  
  final_LCCs[[i]] = out %<>% bind_rows()
}

final_LCCs %<>% bind_rows()
```

# Adj p-values
```{r}
data <- final_LCCs

get_total_genes <- autoimmune_LCC_df[,c("Disease","num_genes")]
colnames(get_total_genes)[1] <- "Disease_Name"

data <- full_join(data, get_total_genes)

data %<>%
  mutate(padj = p.adjust(LCC_p, method = "fdr")) %>%
  mutate(Significance = ifelse(padj < 0.05, "padj < 0.05", "padj > 0.05")) %>%
  mutate(rLCC = (LCC_size/Disease_Cell_Genes)*100) %>%
  mutate(rLCC_50 = ifelse(rLCC > 50, "big rLCC", "small rLCC")) %>%
  mutate(shape = ifelse(padj < 0.05, 1, 4))

write.table(data, "/Users/ursulawidocki/Desktop/BarabasiLab/Data/Disease_ImmuneCell_LCCs_newData2_Q10.tsv", row.names = F, col.names = T, sep = "\t")
```

# Make the plots
```{r}
p1 <- data %>%
  filter(Disease_Name %in% signif_dis) %>%
  ggplot() +
  aes(x = Disease_Name, 
      y = Cell, 
      color = rLCC_50,
      size = rLCC,
      shape = Significance) +
  geom_point() +
  scale_color_manual(values = c("big rLCC" = "#007CBE", "small rLCC" = "grey80")) +
  #scale_color_distiller(palette = "YlGnBu",
  #                      direction = 1) +
  # add ggnewscale to plot the red Xs
  scale_shape_manual(values = c(16, 1)) +
  scale_size(range = c(0.1, 5)) + 
  labs(x = "Autoimmune Disease") +
  theme_minimal() +
  theme(axis.text.x = element_text(angle = 90), 
        legend.position = "bottom", 
        panel.grid.major = element_blank(), 
        panel.grid.minor = element_blank(),
        text = element_text(size = 20))

# Disease bar plot
temp <- data %>%
  filter(Disease_Name %in% signif_dis) %>% 
  filter(padj < 0.05) %>% # uveitis dropped here
  group_by(Disease_Name) %>% 
  dplyr::summarise(n=n())
# add rows for diseases with no modules in cells
temp <- rbind(temp, c("narcolepsy", 0)) %>% as.data.frame()
temp$n <- as.integer(temp$n)

p3 <- temp %>%
  ggplot() +
  aes(x = Disease_Name, y = n) +
  geom_bar(fill = "#122265", stat = "identity") + #position = position_dodge(preserve = "single")
  theme_minimal() +
  theme(axis.title.x=element_blank(),
        axis.text.x=element_blank(),
        axis.ticks.x=element_blank()) +
  labs(y = "LCC count")


# Immune cell bar plot
p2 <- data %>%
  filter(Disease_Name %in% signif_dis) %>% 
  filter(padj < 0.05) %>% 
  group_by(Cell) %>% 
  dplyr::summarise(n=n()) %>%
  ggplot() +
  aes(x = Cell, weight = n) +
  geom_bar(fill = "#122265") + # "#333366"
  labs(y = "LCC count") +
  theme_minimal() +
  theme(axis.title.y = element_blank(),
        axis.text.y = element_blank(),
        axis.ticks.y = element_blank()) +
  coord_flip()


CairoPDF("~/Desktop/BarabasiLab/Auto_Innate_GDA/Figures/Disease_ImmuneCell_LCC_newData2_Q10.pdf", width = 15, height = 10)
# patchwork things together

((p3 + plot_spacer() + plot_layout(widths = c(5,1))) / ((p1 + p2) + plot_layout(widths = c(5,1))))+ plot_layout(heights = c(1,6))

dev.off()
```




# Q25
```{r}
immune_cell_info <- summary_immune_cell %>% 
  filter(sum >= Q25) %>%
  select(`Blood cell`,`Gene name`)
cells = unique(summary_immune_cell$`Blood cell`)
reps = 1000
pb <- txtProgressBar(min = 0, max = length(cells), style = 3)
final_LCCs = list()

for(i in 1:length(cells)){
  setTxtProgressBar(pb, i)
  cell =  cells[i]
  genes_in_cell = immune_cell_info %>% 
    filter(`Blood cell` == cell) %>% 
    filter(`Gene name` %in% V(ppi_g)$name) %>% 
    pull(`Gene name`)
  
  n_genes_cell = length(genes_in_cell)
  
  cell_ppi = induced_subgraph(ppi_g, genes_in_cell)
  
  out = list()
  for(j in 1:length(signif_dis)){
    genes_disease = GDA_auto %>%
      filter(NewName %in% signif_dis[j]) %>%
      filter(HGNC_Symbol %in% V(cell_ppi)$name)
    n_genes_disease_cell  = nrow(genes_disease)
    
    if(n_genes_disease_cell > 0){
      LCC_disease_cell= NetSci::LCC_Significance(N = reps, 
                                                    Targets = genes_disease$HGNC_Symbol, 
                                                    G = cell_ppi,
                                                    bins = 1
      )
      
      # store data
      tmp = data.frame(Disease_Name = signif_dis[j], 
                       Cell = cell, 
                       Cell_Genes = n_genes_cell, 
                       Disease_Cell_Genes = n_genes_disease_cell, 
                       
                       LCC_size = LCC_disease_cell$LCC, 
                       LCC_p = LCC_disease_cell$emp_p)
      
      out[[j]] = tmp
    } else {
      
      tmp = data.frame(Disease_Name = signif_dis[j], 
                       Cell = cell, 
                       Cell_Genes = n_genes_cell, 
                       Disease_Cell_Genes = n_genes_disease_cell, 
                       
                       LCC_size = 0, 
                       LCC_p = 1)
      
      out[[j]] = tmp
    }
  }  
  final_LCCs[[i]] = out %<>% bind_rows()
}

final_LCCs %<>% bind_rows()
```

# Adj p-values
```{r}
data <- final_LCCs

get_total_genes <- autoimmune_LCC_df[,c("Disease","num_genes")]
colnames(get_total_genes)[1] <- "Disease_Name"

data <- full_join(data, get_total_genes)

data %<>%
  mutate(padj = p.adjust(LCC_p, method = "fdr")) %>%
  mutate(Significance = ifelse(padj < 0.05, "padj < 0.05", "padj > 0.05")) %>%
  mutate(rLCC = (LCC_size/Disease_Cell_Genes)*100) %>%
  mutate(rLCC_50 = ifelse(rLCC > 50, "big rLCC", "small rLCC")) %>%
  mutate(shape = ifelse(padj < 0.05, 1, 4))

write.table(data, "/Users/ursulawidocki/Desktop/BarabasiLab/Data/Disease_ImmuneCell_LCCs_newData2_Q25.tsv", row.names = F, col.names = T, sep = "\t")
```

# Make the plots
```{r}
p1 <- data %>%
  filter(Disease_Name %in% signif_dis) %>%
  ggplot() +
  aes(x = Disease_Name, 
      y = Cell, 
      color = rLCC_50,
      size = rLCC,
      shape = Significance) +
  geom_point() +
  scale_color_manual(values = c("big rLCC" = "#007CBE", "small rLCC" = "grey80")) +
  #scale_color_distiller(palette = "YlGnBu",
  #                      direction = 1) +
  # add ggnewscale to plot the red Xs
  scale_shape_manual(values = c(16, 1)) +
  scale_size(range = c(0.1, 5)) + 
  labs(x = "Autoimmune Disease") +
  theme_minimal() +
  theme(axis.text.x = element_text(angle = 90), 
        legend.position = "bottom", 
        panel.grid.major = element_blank(), 
        panel.grid.minor = element_blank(),
        text = element_text(size = 20))

# Disease bar plot
temp <- data %>%
  filter(Disease_Name %in% signif_dis) %>% 
  filter(padj < 0.05) %>% # uveitis dropped here
  group_by(Disease_Name) %>% 
  dplyr::summarise(n=n())
# add rows for diseases with no modules in cells
temp <- rbind(temp, c("narcolepsy", 0)) %>% as.data.frame()
temp$n <- as.integer(temp$n)

p3 <- temp %>%
  ggplot() +
  aes(x = Disease_Name, y = n) +
  geom_bar(fill = "#122265", stat = "identity") + #position = position_dodge(preserve = "single")
  theme_minimal() +
  theme(axis.title.x=element_blank(),
        axis.text.x=element_blank(),
        axis.ticks.x=element_blank()) +
  labs(y = "LCC count")


# Immune cell bar plot
p2 <- data %>%
  filter(Disease_Name %in% signif_dis) %>% 
  filter(padj < 0.05) %>% 
  group_by(Cell) %>% 
  dplyr::summarise(n=n()) %>%
  ggplot() +
  aes(x = Cell, weight = n) +
  geom_bar(fill = "#122265") + # "#333366"
  labs(y = "LCC count") +
  theme_minimal() +
  theme(axis.title.y = element_blank(),
        axis.text.y = element_blank(),
        axis.ticks.y = element_blank()) +
  coord_flip()


CairoPDF("~/Desktop/BarabasiLab/Auto_Innate_GDA/Figures/Disease_ImmuneCell_LCC_newData2_Q25.pdf", width = 15, height = 10)
# patchwork things together

((p3 + plot_spacer() + plot_layout(widths = c(5,1))) / ((p1 + p2) + plot_layout(widths = c(5,1))))+ plot_layout(heights = c(1,6))

dev.off()
```







# Q75
```{r}
immune_cell_info <- summary_immune_cell %>% 
  filter(sum >= Q75) %>%
  select(`Blood cell`,`Gene name`)
cells = unique(summary_immune_cell$`Blood cell`)
reps = 1000
pb <- txtProgressBar(min = 0, max = length(cells), style = 3)
final_LCCs = list()

for(i in 1:length(cells)){
  setTxtProgressBar(pb, i)
  cell =  cells[i]
  genes_in_cell = immune_cell_info %>% 
    filter(`Blood cell` == cell) %>% 
    filter(`Gene name` %in% V(ppi_g)$name) %>% 
    pull(`Gene name`)
  
  n_genes_cell = length(genes_in_cell)
  
  cell_ppi = induced_subgraph(ppi_g, genes_in_cell)
  
  out = list()
  for(j in 1:length(signif_dis)){
    genes_disease = GDA_auto %>%
      filter(NewName %in% signif_dis[j]) %>%
      filter(HGNC_Symbol %in% V(cell_ppi)$name)
    n_genes_disease_cell  = nrow(genes_disease)
    
    if(n_genes_disease_cell > 0){
      LCC_disease_cell= NetSci::LCC_Significance(N = reps, 
                                                    Targets = genes_disease$HGNC_Symbol, 
                                                    G = cell_ppi,
                                                    bins = 1
      )
      
      # store data
      tmp = data.frame(Disease_Name = signif_dis[j], 
                       Cell = cell, 
                       Cell_Genes = n_genes_cell, 
                       Disease_Cell_Genes = n_genes_disease_cell, 
                       
                       LCC_size = LCC_disease_cell$LCC, 
                       LCC_p = LCC_disease_cell$emp_p)
      
      out[[j]] = tmp
    } else {
      
      tmp = data.frame(Disease_Name = signif_dis[j], 
                       Cell = cell, 
                       Cell_Genes = n_genes_cell, 
                       Disease_Cell_Genes = n_genes_disease_cell, 
                       
                       LCC_size = 0, 
                       LCC_p = 1)
      
      out[[j]] = tmp
    }
  }  
  final_LCCs[[i]] = out %<>% bind_rows()
}

final_LCCs %<>% bind_rows()
```

# Adj p-values
```{r}
data <- final_LCCs

get_total_genes <- autoimmune_LCC_df[,c("Disease","num_genes")]
colnames(get_total_genes)[1] <- "Disease_Name"

data <- full_join(data, get_total_genes)

data %<>%
  mutate(padj = p.adjust(LCC_p, method = "fdr")) %>%
  mutate(Significance = ifelse(padj < 0.05, "padj < 0.05", "padj > 0.05")) %>%
  mutate(rLCC = (LCC_size/Disease_Cell_Genes)*100) %>%
  mutate(rLCC_50 = ifelse(rLCC > 50, "big rLCC", "small rLCC")) %>%
  mutate(shape = ifelse(padj < 0.05, 1, 4))

write.table(data, "/Users/ursulawidocki/Desktop/BarabasiLab/Data/Disease_ImmuneCell_LCCs_newData2_Q75.tsv", row.names = F, col.names = T, sep = "\t")
```

# Make the plots
```{r}
p1 <- data %>%
  filter(Disease_Name %in% signif_dis) %>%
  ggplot() +
  aes(x = Disease_Name, 
      y = Cell, 
      color = rLCC_50,
      size = rLCC,
      shape = Significance) +
  geom_point() +
  scale_color_manual(values = c("big rLCC" = "#007CBE", "small rLCC" = "grey80")) +
  #scale_color_distiller(palette = "YlGnBu",
  #                      direction = 1) +
  # add ggnewscale to plot the red Xs
  scale_shape_manual(values = c(16, 1)) +
  scale_size(range = c(0.1, 5)) + 
  labs(x = "Autoimmune Disease") +
  theme_minimal() +
  theme(axis.text.x = element_text(angle = 90), 
        legend.position = "bottom", 
        panel.grid.major = element_blank(), 
        panel.grid.minor = element_blank(),
        text = element_text(size = 20))

# Disease bar plot
temp <- data %>%
  filter(Disease_Name %in% signif_dis) %>% 
  filter(padj < 0.05) %>% # uveitis dropped here
  group_by(Disease_Name) %>% 
  dplyr::summarise(n=n())
# add rows for diseases with no modules in cells
temp <- rbind(temp, c("narcolepsy", 0)) %>% as.data.frame()
temp$n <- as.integer(temp$n)

p3 <- temp %>%
  ggplot() +
  aes(x = Disease_Name, y = n) +
  geom_bar(fill = "#122265", stat = "identity") + #position = position_dodge(preserve = "single")
  theme_minimal() +
  theme(axis.title.x=element_blank(),
        axis.text.x=element_blank(),
        axis.ticks.x=element_blank()) +
  labs(y = "LCC count")


# Immune cell bar plot
p2 <- data %>%
  filter(Disease_Name %in% signif_dis) %>% 
  filter(padj < 0.05) %>% 
  group_by(Cell) %>% 
  dplyr::summarise(n=n()) %>%
  ggplot() +
  aes(x = Cell, weight = n) +
  geom_bar(fill = "#122265") + # "#333366"
  labs(y = "LCC count") +
  theme_minimal() +
  theme(axis.title.y = element_blank(),
        axis.text.y = element_blank(),
        axis.ticks.y = element_blank()) +
  coord_flip()


CairoPDF("~/Desktop/BarabasiLab/Auto_Innate_GDA/Figures/Disease_ImmuneCell_LCC_newData2_Q75.pdf", width = 15, height = 10)
# patchwork things together

((p3 + plot_spacer() + plot_layout(widths = c(5,1))) / ((p1 + p2) + plot_layout(widths = c(5,1))))+ plot_layout(heights = c(1,6))

dev.off()
```







# Q90
```{r}
immune_cell_info <- summary_immune_cell %>% 
  filter(sum >= Q90) %>%
  select(`Blood cell`,`Gene name`)
cells = unique(summary_immune_cell$`Blood cell`)
reps = 1000
pb <- txtProgressBar(min = 0, max = length(cells), style = 3)
final_LCCs = list()

for(i in 1:length(cells)){
  setTxtProgressBar(pb, i)
  cell =  cells[i]
  genes_in_cell = immune_cell_info %>% 
    filter(`Blood cell` == cell) %>% 
    filter(`Gene name` %in% V(ppi_g)$name) %>% 
    pull(`Gene name`)
  
  n_genes_cell = length(genes_in_cell)
  
  cell_ppi = induced_subgraph(ppi_g, genes_in_cell)
  
  out = list()
  for(j in 1:length(signif_dis)){
    genes_disease = GDA_auto %>%
      filter(NewName %in% signif_dis[j]) %>%
      filter(HGNC_Symbol %in% V(cell_ppi)$name)
    n_genes_disease_cell  = nrow(genes_disease)
    
    if(n_genes_disease_cell > 0){
      LCC_disease_cell= NetSci::LCC_Significance(N = reps, 
                                                    Targets = genes_disease$HGNC_Symbol, 
                                                    G = cell_ppi,
                                                    bins = 1
      )
      
      # store data
      tmp = data.frame(Disease_Name = signif_dis[j], 
                       Cell = cell, 
                       Cell_Genes = n_genes_cell, 
                       Disease_Cell_Genes = n_genes_disease_cell, 
                       
                       LCC_size = LCC_disease_cell$LCC, 
                       LCC_p = LCC_disease_cell$emp_p)
      
      out[[j]] = tmp
    } else {
      
      tmp = data.frame(Disease_Name = signif_dis[j], 
                       Cell = cell, 
                       Cell_Genes = n_genes_cell, 
                       Disease_Cell_Genes = n_genes_disease_cell, 
                       
                       LCC_size = 0, 
                       LCC_p = 1)
      
      out[[j]] = tmp
    }
  }  
  final_LCCs[[i]] = out %<>% bind_rows()
}

final_LCCs %<>% bind_rows()
```

# Adj p-values
```{r}
data <- final_LCCs

get_total_genes <- autoimmune_LCC_df[,c("Disease","num_genes")]
colnames(get_total_genes)[1] <- "Disease_Name"

data <- full_join(data, get_total_genes)

data %<>%
  mutate(padj = p.adjust(LCC_p, method = "fdr")) %>%
  mutate(Significance = ifelse(padj < 0.05, "padj < 0.05", "padj > 0.05")) %>%
  mutate(rLCC = (LCC_size/Disease_Cell_Genes)*100) %>%
  mutate(rLCC_50 = ifelse(rLCC > 50, "big rLCC", "small rLCC")) %>%
  mutate(shape = ifelse(padj < 0.05, 1, 4))

write.table(data, "/Users/ursulawidocki/Desktop/BarabasiLab/Data/Disease_ImmuneCell_LCCs_newData2_Q90.tsv", row.names = F, col.names = T, sep = "\t")
```

# Make the plots
```{r}
p1 <- data %>%
  filter(Disease_Name %in% signif_dis) %>%
  ggplot() +
  aes(x = Disease_Name, 
      y = Cell, 
      color = rLCC_50,
      size = rLCC,
      shape = Significance) +
  geom_point() +
  scale_color_manual(values = c("big rLCC" = "#007CBE", "small rLCC" = "grey80")) +
  #scale_color_distiller(palette = "YlGnBu",
  #                      direction = 1) +
  # add ggnewscale to plot the red Xs
  scale_shape_manual(values = c(16, 1)) +
  scale_size(range = c(0.1, 5)) + 
  labs(x = "Autoimmune Disease") +
  theme_minimal() +
  theme(axis.text.x = element_text(angle = 90), 
        legend.position = "bottom", 
        panel.grid.major = element_blank(), 
        panel.grid.minor = element_blank(),
        text = element_text(size = 20))

# Disease bar plot
temp <- data %>%
  filter(Disease_Name %in% signif_dis) %>% 
  filter(padj < 0.05) %>% # uveitis dropped here
  group_by(Disease_Name) %>% 
  dplyr::summarise(n=n())
# add rows for diseases with no modules in cells
temp <- rbind(temp, c("narcolepsy", 0)) %>% as.data.frame()
temp <- rbind(temp, c("uveitis", 0)) %>% as.data.frame()
temp <- rbind(temp, c("scleroderma", 0)) %>% as.data.frame()
temp <- rbind(temp, c("psoriasis", 0)) %>% as.data.frame()
temp <- rbind(temp, c("multiple sclerosis", 0)) %>% as.data.frame()
temp <- rbind(temp, c("lupus erythmatosus systemic", 0)) %>% as.data.frame()
temp <- rbind(temp, c("idiopathic pulmonary fibrosis", 0)) %>% as.data.frame()
temp <- rbind(temp, c("glomerulonephritis", 0)) %>% as.data.frame()
temp <- rbind(temp, c("alopecia areata", 0)) %>% as.data.frame()
temp <- rbind(temp, c("amyloidosis", 0)) %>% as.data.frame()
temp <- rbind(temp, c("anemia hemolytic", 0)) %>% as.data.frame()
temp <- rbind(temp, c("arthritis juvenile", 0)) %>% as.data.frame()
temp <- rbind(temp, c("arthritis rheumatoid", 0)) %>% as.data.frame()
temp <- rbind(temp, c("colitis ulcerative", 0)) %>% as.data.frame()
temp <- rbind(temp, c("crohn disease", 0)) %>% as.data.frame()
temp <- rbind(temp, c("diabetes mellitus type 1", 0)) %>% as.data.frame()
temp <- rbind(temp, c("endometriosis", 0)) %>% as.data.frame()
temp$n <- as.integer(temp$n)

p3 <- temp %>%
  ggplot() +
  aes(x = Disease_Name, y = n) +
  geom_bar(fill = "#122265", stat = "identity") + #position = position_dodge(preserve = "single")
  theme_minimal() +
  theme(axis.title.x=element_blank(),
        axis.text.x=element_blank(),
        axis.ticks.x=element_blank()) +
  labs(y = "LCC count")


# Immune cell bar plot
p2 <- data %>%
  filter(Disease_Name %in% signif_dis) %>% 
  filter(padj < 0.05) %>% 
  group_by(Cell) %>% 
  dplyr::summarise(n=n()) %>%
  ggplot() +
  aes(x = Cell, weight = n) +
  geom_bar(fill = "#122265") + # "#333366"
  labs(y = "LCC count") +
  theme_minimal() +
  theme(axis.title.y = element_blank(),
        axis.text.y = element_blank(),
        axis.ticks.y = element_blank()) +
  coord_flip()


CairoPDF("~/Desktop/BarabasiLab/Auto_Innate_GDA/Figures/Disease_ImmuneCell_LCC_newData2_Q90.pdf", width = 15, height = 10)
# patchwork things together

((p3 + plot_spacer() + plot_layout(widths = c(5,1))) / ((p1 + p2) + plot_layout(widths = c(5,1))))+ plot_layout(heights = c(1,6))

dev.off()
```




# Old Code ######################################################################################
######################### Code I'm not using
# Make a sankey plot


library(ggsankey)

data %<>% arrange(prop_lcc)
data$Disease = factor(row.names(data), 
                                   levels = row.names(data))
data_long <- make_long(data, Disease_Name, Tissue, value = prop_lcc)

ggplot(data_long, aes(x = x, 
               next_x = next_x, 
               node = node, 
               next_node = next_node,
               fill = factor(node),
               label = node)) +
  geom_sankey(aes(width = value), flow.alpha = .6,
              node.color = "gray30") +
  geom_sankey_label(size = 3, color = "white", fill = "gray40") +
  labs(x = NULL) +
  theme_sankey(base_size = 16) +
  theme(legend.position = "none")

library(UpSetR)
tissue_dat$term <- gsub("SMTSD", "", tissue_dat$term) 

CairoPDF("/Users/ursulawidocki/Desktop/BarabasiLab/Auto_Innate_GDA/Figures/logNorm_DE_Tissue_genes_UpSet.pdf", 
         width = 35, height = 25)

tissue_dat %>%
   filter(p_adj < 0.05) %>%
   select(HGNC_Symbol, term) %>%
   mutate(term = stringr::str_remove(term, "tissue")) %>%
   ungroup() %>% 
   pivot_wider(names_from = term, 
               values_from = term, 
               values_fn = length, 
               values_fill = 0) %>%
   as.data.frame() %>% 
   UpSetR::upset(.,nsets = 56, 
                 order.by =  "freq",
                 text.scale = 3, point.size=3)

dev.off()







# make Sankey plot... but cannot save as pdf
library(networkD3)

brain <- c("Brain Amygdala","Brain Anterior cingulate cortex BA24","Brain Caudate basal ganglia","Brain Cerebellar Hemisphere",
           "Brain Cerebellum" ,"Brain Cortex" ,"Brain Frontal Cortex BA9","Brain Hippocampus" , "Brain Hypothalamus",                    "Brain Nucleus accumbens basal ganglia", "Brain Putamen basal ganglia" ,"Brain Spinal cord cervical c 1", "Brain Substantia nigra")

data %<>% arrange(LCC_size)
data$Disease = factor(row.names(data), 
                                   levels = row.names(data))

g_tiss_dis <- data %>%
  filter(padj < 0.05) %>%
  select(c("Disease_Name", "Tissue", "LCC_size")) %>%
  filter(Tissue %ni% brain) %>%
  graph.data.frame(., directed = F)
tiss_dis_matrix <- get.adjacency(g_tiss_dis, sparse = FALSE, attr='LCC_size')
g_tiss_dis = graph_from_adjacency_matrix(tiss_dis_matrix, mode = "undirected", weighted = T)




clust <- cluster_walktrap(g_tiss_dis)
members <- membership(clust)
g_tiss_dis_D3 <- igraph_to_networkD3(g_tiss_dis, group = members)
myColors <-rep(c("tomato3","green"), times=c(21,38))
# sankeyNetwork()
sankeyNetwork(Links = g_tiss_dis_D3$links, Nodes = g_tiss_dis_D3$nodes, Source = "source",
             Target = "target", Value = "value", NodeID = "name",
             colourScale = myColors,
             fontSize = 18, nodeWidth = 30)

URL <- paste0(
        "https://cdn.rawgit.com/christophergandrud/networkD3/",
        "master/JSONdata/energy.json")
Energy <- jsonlite::fromJSON(URL)
# Plot
sankeyNetwork(Links = Energy$links, Nodes = Energy$nodes, Source = "source",
             Target = "target", Value = "value", NodeID = "name",
             units = "TWh", fontSize = 12, nodeWidth = 30)

# filter out genes that are not assoc with any diseases
#tissue_dat <- tissue_dat %>% select(-Name) %>% filter(Description %in% V(ppi_g)$name)
# if there is a duplicate, take the sum of their collective expression values
summary_gtex = tissue_dat %>% 
  pivot_longer(-c(1)) %>%
  #filter(value > 0) %>% 
  group_by(name, Description) %>%
  summarise(sum = sum(value)) %>% 
  mutate(Q50 = quantile(sum, 0.50))
# Clean up tissue names in data
summary_gtex$name = summary_gtex$name %>%
  stringr::str_replace_all(., "\\.", " ") %>%
  stringr::str_squish()
summary_gtex %<>% filter(name %ni% c("Cells EBV transformed lymphocytes", "Cells Cultured fibroblasts", "Nerve Tibial"))
